basic crud operation and thirdparty api integration

+++ Crud_API/Infrastructure/Database/Models/StudentCourse.cs
@@ -5,8 +5,8 @@ namespace Crud_API.Infrastructure.Database.Models
     public class StudentCourse : IEntity
     {
         public long Id { get; set; }
-        public int StudentId { get; set; }
-        public int CourseId { get; set; }
+        public long StudentId { get; set; }
+        public long CourseId { get; set; }
 
         #region Navigation properties
